Migração de EF Core para Dapper e remoção do AutoMapper

+++ EasyBudget/EasyBudget/Data/Dto/CategoryDto/ReplaceCategoryDto.cs
@@ -2,6 +2,6 @@
 {
     public class ReplaceCategoryDto
     {
-        public long ReplaceCategoryId { get; set; }
+        public int ReplaceCategoryId { get; set; }
     }
 }
